requirements.txt
anyio==3.6.2
attrs==22.2.0
blinker==1.5
Brotli==1.0.9
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crit==3.17.1
cryptography==38.0.4
dbus-python==1.3.2
distro==1.8.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fuse-python==1.0.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
idna==3.3
jsonpointer==2.3
jsonschema==4.10.3
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
Mako==1.2.4.dev0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
meson==1.0.1
msgpack==1.0.3
mutagen==1.46.0
nala==0.12.2
oauthlib==3.2.2
pexpect==4.8.0
protobuf==4.21.12
ptyprocess==0.7.0
pycryptodomex==3.11.0
pycurl==7.45.2
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pylibacl==0.7.0
pynvim==0.4.2
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
PySimpleSOAP==1.16.2
python-apt==2.6.0
python-debian==0.1.49
python-debianbts==4.0.1
python-dotenv==0.21.0
pyxattr==0.8.1
PyYAML==6.0
reportbug==12.0.0
requests==2.28.1
rfc3986==1.5.0
rfc3987==1.3.8
rich==13.3.1
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
texttable==1.6.7
tomli==2.0.1
tornado==6.2
typer==0.7.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wadllib==1.3.6
webcolors==1.11.1
websocket-client==1.2.3
websockets==10.4
yt-dlp==2023.3.4
